Move tokenfactory genesis logic into keeper package, de-expose setAuthorityMetadata
