+++ x/tokenfactory/keeper/admins.go
@@ -19,8 +19,8 @@ func (k Keeper) GetAuthorityMetadata(ctx sdk.Context, denom string) (types.Denom
 	return metadata, nil
 }
 
-// SetAuthorityMetadata stores authority metadata for a specific denom
-func (k Keeper) SetAuthorityMetadata(ctx sdk.Context, denom string, metadata types.DenomAuthorityMetadata) error {
+// setAuthorityMetadata stores authority metadata for a specific denom
+func (k Keeper) setAuthorityMetadata(ctx sdk.Context, denom string, metadata types.DenomAuthorityMetadata) error {
 	err := metadata.Validate()
 	if err != nil {
 		return err
@@ -45,5 +45,5 @@ func (k Keeper) setAdmin(ctx sdk.Context, denom string, admin string) error {
 
 	metadata.Admin = admin
 
-	return k.SetAuthorityMetadata(ctx, denom, metadata)
+	return k.setAuthorityMetadata(ctx, denom, metadata)
 }
